Add a seperate footer
